Fix fox and tracking orbit
