+++ util.py
@@ -4,14 +4,9 @@ import os
 from typing import Any, Optional, Tuple
 
 import cv2
-import json
 import numpy as np
 import requests
 from tqdm import tqdm
-from typing import TypedDict, List
-import time
-from dataclasses import dataclass
-from datetime import datetime
 
 
 def download_file(path: str, link: str) -> None:
@@ -145,191 +140,3 @@ class HostSync(object):
                     del self.dict[rm]
                 return ret
         return None
-
-
-@dataclass
-class PosLog:
-    time: int
-    x: float
-    y: float
-    z: float
-
-
-class LogJson(TypedDict):
-    id: int
-    name: str
-    time: float
-    pos: List[Tuple[float, float, float]]
-
-
-class OrbitData(object):
-    def __init__(self, name: str, id: int, pos_log: PosLog):
-        self.name: str = name
-        self.id: int = id
-        self.pos_log: List[PosLog] = []
-        self.tmp_pos_log: List[PosLog] = [pos_log]
-
-
-class OrbitDataList(object):
-    def __init__(self, labels: List[str], log_path: Optional[str] = None):
-        self.LOGGING_INTEREVAL = 0.5
-        self.start_time = time.time()
-        self.last_update_time = 0.0
-        self.data: List[OrbitData] = []
-        self.labels: List[str] = labels
-        if log_path is not None:
-            if not os.path.exists(log_path):
-                os.makedirs(log_path)
-            current_time = datetime.now()
-            self.file_name = (
-                log_path + f"/data_{current_time.strftime('%Y%m%d_%H%M%S')}.json"
-            )
-            init_json = {}
-            init_json["start_time"] = f"{current_time.strftime('%Y%m%d_%H%M%S')}"
-            init_json["interval"] = self.LOGGING_INTEREVAL
-            init_json["logs"] = []
-            with open(self.file_name, mode="wt", encoding="utf-8") as f:
-                json.dump(init_json, f, ensure_ascii=False, indent=2)
-
-    def __del__(self):
-        for data in self.data:
-            self.save_pos_log(data)
-
-    def get_cur_time(self) -> float:
-        return time.time() - self.start_time
-
-    def get_orbit_from_id(self, id: int) -> OrbitData:
-        for data in self.data:
-            if data.id == id:
-                return data
-        return None
-
-    def add_new_data(self, tracklet: Any):
-        pos_data = PosLog(
-            time=self.get_cur_time(),
-            x=tracklet.spatialCoordinates.x,
-            y=tracklet.spatialCoordinates.y,
-            z=tracklet.spatialCoordinates.z,
-        )
-        self.data.append(
-            OrbitData(
-                name=self.labels[tracklet.label], id=tracklet.id, pos_log=pos_data
-            )
-        )
-
-    def add_track_data(self, tracklet: Any, pos_list: List[PosLog]):
-        pos_data = PosLog(
-            time=self.get_cur_time(),
-            x=tracklet.spatialCoordinates.x,
-            y=tracklet.spatialCoordinates.y,
-            z=tracklet.spatialCoordinates.z,
-        )
-        pos_list.append(pos_data)
-
-    def update_orbit_data(self, tracklets: List[Any]) -> None:
-        if tracklets is None:
-            return
-        for tracklet in tracklets:
-            new_data = True
-            for data in self.data:
-                if tracklet.id == data.id:
-                    self.add_track_data(tracklet, data.tmp_pos_log)
-                    new_data = False
-            if new_data:
-                self.add_new_data(tracklet)
-        self.fix_pos_log()
-        self.remove_old_data(tracklets)
-
-    def weighted_average_position(
-        self, pos_logs: List[PosLog], target_time: float
-    ) -> PosLog:
-        """
-        指定された時間での位置を、重み付け平均から推測する。
-
-        Args:
-            pos_logs (List[PosLog]): 時間と位置のログのリスト。
-            target_time (float): 推測したい時間。
-
-        Returns:
-            PosLog: 推測された位置。
-        """
-        total_weight = 0.0
-        weighted_sum_x = 0.0
-        weighted_sum_y = 0.0
-        weighted_sum_z = 0.0
-        for log in pos_logs:
-            # 時間差に基づいた重みを計算（時間差が小さいほど重みが大きくなる）
-            weight = 1 / (
-                abs(log.time - target_time) + 1e-9
-            )  # ゼロ除算を避けるための微小値
-            total_weight += weight
-            weighted_sum_x += log.x * weight
-            weighted_sum_y += log.y * weight
-            weighted_sum_z += log.z * weight
-        # 重み付け平均を計算
-        avg_x = weighted_sum_x / total_weight
-        avg_y = weighted_sum_y / total_weight
-        avg_z = weighted_sum_z / total_weight
-
-        return PosLog(time=target_time, x=avg_x, y=avg_y, z=avg_z)
-
-    def fix_pos_log(self) -> None:
-        cur_time = self.get_cur_time()
-        while True:
-            next_time = self.last_update_time + self.LOGGING_INTEREVAL * 3 / 2
-            if cur_time - next_time < 0:
-                break
-            for data in self.data:
-                tmp_list: List[PosLog] = []
-                while True:
-                    if len(data.tmp_pos_log) > 0:
-                        if data.tmp_pos_log[0].time < next_time:
-                            tmp_list.append(data.tmp_pos_log.pop(0))
-                        else:
-                            break
-                    else:
-                        break
-                if len(tmp_list) > 0:
-                    data.pos_log.append(
-                        self.weighted_average_position(
-                            tmp_list, self.last_update_time + self.LOGGING_INTEREVAL
-                        )
-                    )
-            self.last_update_time += self.LOGGING_INTEREVAL
-
-    def remove_old_data(self, tracklets: List[Any]) -> None:
-        new_data = []
-        for data in self.data:
-            is_tracking = False
-            for tracklet in tracklets:
-                if tracklet.id == data.id:
-                    is_tracking = True
-                    break
-            if not is_tracking:
-                if self.file_name is not None:
-                    self.save_pos_log(data)
-            else:
-                new_data.append(data)
-        self.data = new_data
-
-    def save_pos_log(self, data: OrbitData) -> None:
-        if len(data.pos_log) == 0:
-            return
-        new_data: LogJson = {
-            "id": data.id,
-            "name": data.name,
-            "time": data.pos_log[0].time,
-            "pos": [
-                (
-                    round(pos.x / 1000.0, 3),
-                    round(pos.y / 1000.0, 3),
-                    round(pos.z / 1000.0, 3),
-                )
-                for pos in data.pos_log
-            ],
-        }
-        json_open = open(self.file_name, "r")
-        log_file = json.load(json_open)
-        log_file["logs"].append(new_data)
-        with open(self.file_name, mode="wt", encoding="utf-8") as f:
-            json.dump(log_file, f, ensure_ascii=False, indent=2)
